[#136] Reorganize seeds and allow to run them on demand
Split SeedDevelopment to SeedUsers and SeedProducts, which are
independent.
Moreover, allow to turn the seeds on in profiles other than development
by setting appropriate profile `run-seed-<name>`.
